兼容默认值

+++ app/src/main/java/com/zhangzheng/easystore/TestStorage.kt
@@ -4,6 +4,6 @@ import com.zhangzheng.easystore.library.Storable
 
 interface TestStorage :Storable{
     var name:String
-    var count:Int
+    var count:Float
     var isBool:Boolean
 }
